+++ jekyll/version-history.markdown
@@ -4,6 +4,16 @@ permalink: /version-history/
 published: true
 ---
 
+### v1.6.0
+
+ - add pages for Aliases and Establishments and their searching thereof
+ - make the search boxes bigger and disable them until the data is loaded
+ - move the site's title and logo to the top center (away from the left)
+ - give the navigation menu a new layout
+ - better the layout for the tracks list of each album on the "Track Review" page
+
+---
+
 ### v1.5.4
 
  - fix the index page layout by creating a redirect to another page